+++ test/resources/date-doc-definitions.js
@@ -17,6 +17,10 @@
       },
       formatValidationProp: {
         type: 'date'
+      },
+      immutableValidationProp: {
+        type: 'date',
+        immutableWhenSet: true
       }
     }
   },