+++ test/resources/immutable-when-set-strict-doc-definitions.js
@@ -9,7 +9,27 @@
         type: 'string',
         immutableWhenSetStrict: true
       },
-      dynamicPropertiesAreImmutable: {
+      staticImmutableDateProp: {
+        type: 'date',
+        immutableStrict: true
+      },
+      staticImmutableDatetimeProp: {
+        type: 'datetime',
+        immutableStrict: true
+      },
+      staticImmutableTimeProp: {
+        type: 'time',
+        immutableStrict: true
+      },
+      staticImmutableTimezoneProp: {
+        type: 'timezone',
+        immutableStrict: true
+      },
+      staticImmutableUuidProp: {
+        type: 'uuid',
+        immutableStrict: true
+      },
+    dynamicPropertiesAreImmutable: {
         type: 'boolean'
       },
       dynamicValidationProp: {