+++ test/resources/timezone-doc-definitions.js
@@ -15,6 +15,10 @@
         type: 'timezone',
         minimumValueExclusive: '-1131',
         maximumValueExclusive: '+12:31'
+      },
+      immutableValidationProp: {
+        type: 'timezone',
+        immutable: true
       }
     }
   },